round price to nearest integer in cents.
Prices are entered in format X.XX.   They are then multiplied by 100 to give
number of cents.  Since this is floating point, they're sometimes off slightly,
so dividing by 100 again to get the display value will be wonky.

The solution is to round to nearest int after multiplying so that all prices
are in an integer number of cents.